u converter done! Moving on to xX converters... Note that I changed the ft_putnbr_fd function to be able to print long long int, thus making it in range to print unsigned int needed for the u converter..

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/20 12:24:28 by smagdela          #+#    #+#             */
-/*   Updated: 2021/08/13 12:38:07 by smagdela         ###   ########.fr       */
+/*   Updated: 2021/08/24 22:11:08 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long long int n, int fd);
 void	ft_putnbr_base_fd(int n, int fd, const char *base);
 
 typedef struct s_list
